Fix predicates + load issue

+++ shared/origami/include/origami/utils.hpp
@@ -102,15 +102,15 @@ namespace origami
                                   size_t            MT_K,
                                   bool              print);
 
-        size_t select_best_wgm(const hardware_t& hardware,
-                               size_t            M,
-                               size_t            N,
-                               size_t            K,
-                               size_t            batch,
-                               size_t            MT_M,
-                               size_t            MT_N,
-                               size_t            MT_K,
-                               bool              print);
+        int32_t select_best_wgm(const hardware_t& hardware,
+                                size_t            M,
+                                size_t            N,
+                                size_t            K,
+                                size_t            batch,
+                                size_t            MT_M,
+                                size_t            MT_N,
+                                size_t            MT_K,
+                                bool              print);
 
         double compute_tflops_from_latency(double latency_cycles,
                                            size_t M,
